% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\docType{data}
\name{AGGFNTBL}
\alias{AGGFNTBL}
\title{Table of allowable aggregation functions}
\format{An object of class \code{list} of length 5.}
\usage{
AGGFNTBL
}
\description{
Table of allowable aggregation functions
}
\keyword{internal}
