% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{alternate_filename}
\alias{alternate_filename}
\title{Generate an alternate file name, if input name is already in use.}
\usage{
alternate_filename(name)
}
\arguments{
\item{name}{The intended file name.}
}
\description{
Generate the alternate name by appending NNN to the base name, where NNN is
the smallest integer for which the resulting name isn't already in use.
}
\keyword{internal}
