% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{canonicalForm}
\alias{canonicalForm}
\title{Convert a variable name to canonical form}
\usage{
canonicalForm(var)
}
\arguments{
\item{var}{Variable name to convert}
}
\description{
Canonical form means all non-alphanumerics are converted to \code{_}, and
"module." is prepended to the name.
}
\keyword{internal}
