% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{clamp}
\alias{clamp}
\title{Clamp a vector of values to a specified range.}
\usage{
clamp(x, xlo, xhi)
}
\arguments{
\item{x}{Vector of values to clamp}

\item{xlo}{Lower end of the range}

\item{xhi}{Upper end of the range}
}
\description{
How is this not a built-in function already?
}
\keyword{internal}
