% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{compsplt}
\alias{compsplt}
\title{Split composite units}
\usage{
compsplt(unit, convmat)
}
\arguments{
\item{unit}{Either iunit or ounit (see modules above)}

\item{convmat}{Conversion matrices stored in sysdata}
}
\description{
Split composite units (EJ/yr, billion pass-km, etc.), and find canonical version of unit stored
in the corresponding conversion matrix.
All units in these matrices needed to be lower case.
These matrices need to be symmetric for this function to work for both input and output units
}
