% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{dofilter}
\alias{dofilter}
\title{Apply a filter string to a table}
\usage{
dofilter(fvec, tbl)
}
\arguments{
\item{fvec}{Vector of str_match outputs for a single filter}

\item{tbl}{Table to filter}
}
\value{
logical vector
}
\description{
Parse the filter string and compute a logical vector indicating for each row
whether the row is in the result or not.  This logical vector is what is
returned, so that we don't attempt to modify the data set until all of the
filters have been calculated (allowing us to change the data frame just
once).
}
\keyword{internal}
