% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{extractunit}
\alias{extractunit}
\title{Determine which unit in a list a string matches}
\usage{
extractunit(unitlist, unitstr, type)
}
\arguments{
\item{unitlist}{List of allowed units (e.g., the column names of a conversion
matrix).}

\item{unitstr}{The string to match against the allowed units}

\item{type}{The type of unit we were trying to match (used to create warning
messages).}
}
\description{
Match a string one by one against a list of allowed values.  Return the one
that matched, or NULL (with warning) if none of them was matched.
}
\keyword{internal}
