% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter}
\alias{filter}
\title{Filter an output variable}
\usage{
filter(tbl, startyr, endyr, filterstr)
}
\arguments{
\item{tbl}{Table to filter}

\item{startyr}{Start year.  Years less than this will be dropped.}

\item{endyr}{End year.  Years greater than this will be dropped.}

\item{filterstr}{Filter string.  A list of filter s-expressions.  These will
be parsed and applied to the input table.}
}
\value{
Filtered variable table.
}
\description{
Filter an output variable using startyear and endyear values, as well as the
filter string.  The return value is the filtered data set.
}
\keyword{internal}
