% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{filterfn}
\alias{filterfn}
\title{Filtering function}
\usage{
filterfn(module_data, filters)
}
\arguments{
\item{filters}{Character string giving a list of additional filters to be
applied, in s-exp format.}
}
\description{
This function filters query output according to filters parameter
}
\keyword{internal}
