% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\docType{data}
\name{filterfns}
\alias{filterfns}
\title{List of available filter functions}
\format{An object of class \code{list} of length 10.}
\usage{
filterfns
}
\description{
List of available filter functions
}
\keyword{internal}
