% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\docType{data}
\name{filterpattern}
\alias{filterpattern}
\title{Regexp for matching '(arg1; arg2; arg3)'}
\format{An object of class \code{character} of length 1.}
\usage{
filterpattern
}
\description{
Regexp for matching '(arg1; arg2; arg3)'
}
\keyword{internal}
