% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{getaggfn}
\alias{getaggfn}
\title{Look up an aggregation function by name.}
\usage{
getaggfn(fname)
}
\arguments{
\item{fname}{Name of the requested function.}
}
\description{
Look up an aggregation function by name, restricting it to a whitelisted set
of known functions.
}
\keyword{internal}
