% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\docType{data}
\name{GETQ}
\alias{GETQ}
\title{Token indicating get-queries mode for \code{\link{runModule}} family of
functions.}
\format{An object of class \code{character} of length 1.}
\usage{
GETQ
}
\description{
Token indicating get-queries mode for \code{\link{runModule}} family of
functions.
}
\keyword{internal}
