% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{ghg_emissions_by_gwp}
\alias{ghg_emissions_by_gwp}
\title{Helper for GHG Emissions Data Module}
\usage{
ghg_emissions_by_gwp(gwp, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\arguments{
\item{gwp}{Lookup table of ghg to GWP (with units)}
}
\description{
Produce ghg emissions by technology, converted to MTCO2e with given gwp
}
\keyword{internal}
