% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{groupSectors}
\alias{groupSectors}
\title{Group water use sectors into categories}
\usage{
groupSectors(sector)
}
\arguments{
\item{sector}{A column (generally the 'sector' column) from a query result}
}
\description{
Convert the mix of energy, agriculture, and other uses of water into a more
concise grouping. The six larger categories are Forest, Cropland, Biomass,
Other Natural Land, Pasture, and Desert. Any unknown sector passed to this
function will be returned as is.
}
\keyword{internal}
