% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrpt.R
\docType{package}
\name{iamrpt-package}
\alias{iamrpt}
\alias{iamrpt-package}
\title{Convert GCAM output to the format used by public IAM databases}
\description{
Provide functions for converting GCAM output into the format used by most IAM
experiments to enter results into their databases. Users provide a table of
desired outputs, along with options (such as filtering and aggregation), and
the package runs the necessary GCAM queries (no more than once per query) and
passes the results to the functions that produce the output.
}
\section{Usage}{


Coming soon.
}

\author{
\strong{Maintainer}: Robert Link \email{robert.link@pnnl.gov}

Authors:
\itemize{
  \item Xavier Gutierrez \email{xavier.gutierrez@pnnl.gov}
}

}
