% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{iiasa_sortcols}
\alias{iiasa_sortcols}
\title{Put columns in canonical order for IIASA data format}
\usage{
iiasa_sortcols(df)
}
\arguments{
\item{df}{Data frame}
}
\description{
Put columns in canonical order for IIASA data format
}
\keyword{internal}
