% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{iiasafy}
\alias{iiasafy}
\title{Convert a list of tables to a single table in IIASA format}
\usage{
iiasafy(datalist)
}
\arguments{
\item{datalist}{List of data frames, one for each variable.}
}
\description{
The result of this transformation will be a single table with the following
columns:
}
\details{
\itemize{
  \item{Model}
  \item{Scenario}
  \item{Region}
  \item{Variable (taken from the output name of the input)}
  \item{Unit}
  \item{NNNN - one for each year}
}
}
\keyword{internal}
