% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{liquids}
\alias{liquids}
\title{Liquid Fuels}
\usage{
liquids(en, ref)
}
\arguments{
\item{en}{Data returned for final energy query}

\item{ref}{Data returned for refined liquids query}
}
\description{
Calculate biomass liquids energy output from biomass' share of total liquids refining
}
\keyword{internal}
