% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{listModules}
\alias{listModules}
\title{List the variables that the system knows how to generate}
\usage{
listModules()
}
\description{
Returns a character vector listing the variables that currently have recipes
in the system.
}
