% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{mapfuel}
\alias{mapfuel}
\title{Fuel}
\usage{
mapfuel(en, ref = NULL)
}
\arguments{
\item{en}{Data returned for final energy query}

\item{ref}{Data returned for refined liquids query}
}
\description{
Using input col of energy query, create fuel col. Remove input col.
Using bio share of total liquids refining, calculate liquids energy from biomass vs traditional.
All other fuel sources remain unchanged.
}
\keyword{internal}
