% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{merge_scenarios}
\alias{merge_scenarios}
\title{Merge tables for multiple scenarios into a single scenario}
\usage{
merge_scenarios(rawrslts)
}
\arguments{
\item{rawrslts}{Results by scenario and variable passed in from main control
loop}
}
\value{
List of data frames, one for each variable, with all scenarios
included.
}
\description{
For each variable collect the tables for each of the scenarios and fuse them
into a single table.
}
\keyword{internal}
