% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture.R
\name{module.biomass_consumption}
\alias{module.biomass_consumption}
\title{Biomass Consumption Data Module}
\usage{
module.biomass_consumption(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce biomass consumption by crop type
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{output}
  \item{technology}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
