% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture.R
\name{module.biomass_production}
\alias{module.biomass_production}
\title{Biomass Production Data Module}
\usage{
module.biomass_production(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce biomass production by crop type, including forest
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{subsector}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
