% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.building_floorspace}
\alias{module.building_floorspace}
\title{Building Floorspace Module}
\usage{
module.building_floorspace(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit, region, agg_region, add_global)
}
\description{
Produce building floorspace by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
