% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.building_flrsp_cap}
\alias{module.building_flrsp_cap}
\title{Building Floorspace Per Capita Module}
\usage{
module.building_flrsp_cap(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce building floorspace per capita by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
