% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.cement_capita}
\alias{module.cement_capita}
\title{Cement Production Per Capita Module}
\usage{
module.cement_capita(mode, allqueries, aggkeys, aggfn, years, filters, ounit,
  region, agg_region, add_global)
}
\description{
Produce cement production per capita by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
