% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.cement_ghg_emissions_ar4}
\alias{module.cement_ghg_emissions_ar4}
\title{Total Cement GHG Emissions Data Module}
\usage{
module.cement_ghg_emissions_ar4(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit, region, agg_region, add_global)
}
\description{
Produce total cement ghg emissions - includes indirect emissions from electricity and biomass CO2 accounting.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
