% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.cement_production}
\alias{module.cement_production}
\title{Cement Production Module}
\usage{
module.cement_production(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce cement production by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
