% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.co2_emissions_end_use}
\alias{module.co2_emissions_end_use}
\title{CO2 emissions by end-use sector Data Module}
\usage{
module.co2_emissions_end_use(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit, region, agg_region, add_global)
}
\description{
Produce service output by technology and vintage
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
