% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.direct_ghg_emissions_ar4}
\alias{module.direct_ghg_emissions_ar4}
\title{Direct GHG Emissions Data Module}
\usage{
module.direct_ghg_emissions_ar4(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit, region, agg_region, add_global)
}
\description{
Produce direct sector ghg emissions, including CO2 emissions from biomass, converted to MTCO2e with AR4 GWPs
Does not include indirect emissions from electricity consumption.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
