% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.direct_trans_ghg_emissions_ar4}
\alias{module.direct_trans_ghg_emissions_ar4}
\title{Direct Transport GHG Emissions Data Module}
\usage{
module.direct_trans_ghg_emissions_ar4(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce direct transport ghg emissions - includes biomass CO2 accounting.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
