% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity.R
\name{module.electricity}
\alias{module.electricity}
\title{Electricity Data Module}
\usage{
module.electricity(mode, allqueries, aggkeys, aggfn, strtyr, endyr, filters,
  ounit)
}
\description{
Produce electricity by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{subsector}
  \item{technology}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
