% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity.R
\name{module.electricity_shares}
\alias{module.electricity_shares}
\title{Electricity Shares Data Module}
\usage{
module.electricity_shares(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit, region, agg_region, add_global)
}
\description{
Produce electricity shares by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{subsector}
  \item{technology}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
