% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.enteric_fermentation_share}
\alias{module.enteric_fermentation_share}
\title{Share of Ag GHG Emissions from Enteric Fermentation Data Module}
\usage{
module.enteric_fermentation_share(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit, region, agg_region, add_global)
}
\description{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
