% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.final_energy_fuel_shares}
\alias{module.final_energy_fuel_shares}
\title{Final Energy Fuel Shares by Sector Data Module}
\usage{
module.final_energy_fuel_shares(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce final energy fuel shares by sector.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
