% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.final_energy_per_capita}
\alias{module.final_energy_per_capita}
\title{Final Energy per Capita Module}
\usage{
module.final_energy_per_capita(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce total final per capita by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
