% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.final_energy_per_gdp}
\alias{module.final_energy_per_gdp}
\title{Final Energy Per GDP Data Module}
\usage{
module.final_energy_per_gdp(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit, region, agg_region, add_global)
}
\description{
Produce final energy per GDP by region, converted to EJ/$ with AR4 GWPs
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
