% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.final_energy_sector_fuel}
\alias{module.final_energy_sector_fuel}
\title{Final Energy by Sector and Fuel Data Module}
\usage{
module.final_energy_sector_fuel(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce total final by sector and fuel by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
