% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module.gdp_mer_}
\alias{module.gdp_mer_}
\title{GDP MER Data Module}
\usage{
module.gdp_mer_(mode, allqueries, aggkeys, aggfn, years, filters,
  filter_operator, ounit)
}
\description{
Produce GDP MER by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
