% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module.gdp_ppp}
\alias{module.gdp_ppp}
\title{GDP PPP Data Module}
\usage{
module.gdp_ppp(mode, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\description{
Produce GDP PPP by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
