% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.ghg_emissions_intensity_region}
\alias{module.ghg_emissions_intensity_region}
\title{GHG Emissions Intensity By Region Data Module}
\usage{
module.ghg_emissions_intensity_region(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit, region, agg_region, add_global)
}
\description{
Produce ghg emissions intensity (of final energy) by region, converted to MTCO2e/EJ with AR4 GWPs
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
