% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.ghg_emissions_per_capita}
\alias{module.ghg_emissions_per_capita}
\title{GHG Emissions Per Capita Data Module}
\usage{
module.ghg_emissions_per_capita(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce ghg emissions per capita by region, converted to MTCO2e/EJ with AR4 GWPs
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
