% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.ghg_emissions_per_gdp}
\alias{module.ghg_emissions_per_gdp}
\title{GHG Emissions Per GDP Data Module}
\usage{
module.ghg_emissions_per_gdp(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce ghg emissions per GDP by region, converted to MTCO2e/$ with AR4 GWPs
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
