% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.ghg_per_capita_ar4}
\alias{module.ghg_per_capita_ar4}
\title{GHG Emissions Per Capita by sector Data Module}
\usage{
module.ghg_per_capita_ar4(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce ghg emissions per capita by sector - includes indirect emissions from electricity and biomass CO2 accounting.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
