% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.ghg_per_veh_km}
\alias{module.ghg_per_veh_km}
\title{GHG Emissions Per veh-km Data Module}
\usage{
module.ghg_per_veh_km(mode, allqueries, aggkeys, aggfn, years, filters, ounit,
  region, agg_region, add_global)
}
\description{
Produce ghg emissions per veg-km - includes indirect emissions from electricity and biomass CO2 accounting.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
