% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.ghg_trans_en_intensity}
\alias{module.ghg_trans_en_intensity}
\title{GHG Emissions Per Transport Energy Data Module}
\usage{
module.ghg_trans_en_intensity(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit, region, agg_region, add_global)
}
\description{
Produce ghg emissions per transport energy - includes indirect emissions from electricity and biomass CO2 accounting.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
