% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.indirect_ghg_emissions_ar4}
\alias{module.indirect_ghg_emissions_ar4}
\title{Indirect GHG Emissions Data Module}
\usage{
module.indirect_ghg_emissions_ar4(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce indirect sector ghg emissions by sector - emissions from electricity consumption.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
