% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.indirect_trans_ghg_emissions_ar4}
\alias{module.indirect_trans_ghg_emissions_ar4}
\title{Indirect Transport GHG Emissions Data Module}
\usage{
module.indirect_trans_ghg_emissions_ar4(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit, region, agg_region, add_global)
}
\description{
Produce indirect transport ghg emissions - includes indirect emissions from electricity and biomass CO2 accounting in electricity.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
