% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse.R
\name{module.land_cover}
\alias{module.land_cover}
\title{Land Allocation Data Module}
\usage{
module.land_cover(mode, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\description{
Produce land use by region. Creates columns named 'aez land allocation',
'land allocation', and ' land type' representing different levels of
aggregation of the category of land use (the broadest categories being
Biomass, Built-up Area, Cropland Forest, Other Natural Land, Pasture). Note
that the the aggkeys and filters arguments must use these new column names.
}
\details{
The raw table used by this module has columns: \itemize{ \item{scenario}
\item{region} \item{landnode} \item{land_allocation} \item{year} \item{value}
\item{Units} }
}
\keyword{internal}
