% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_use.R
\name{module.liquid_shares}
\alias{module.liquid_shares}
\title{Refined Liquids Shares Module}
\usage{
module.liquid_shares(mode, allqueries, aggkeys, aggfn, years, filters, ounit,
  region, agg_region, add_global)
}
\description{
Produce shares of refined liquids by fuel type by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
