% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{module.load_factors}
\alias{module.load_factors}
\title{Load Factors Data Module}
\usage{
module.load_factors(mode, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\description{
Produce load factors by technology and vintage
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
