% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_energy.R
\name{module.oil_shares}
\alias{module.oil_shares}
\title{Oil Production Type Share Data Module}
\usage{
module.oil_shares(mode, allqueries, aggkeys, aggfn, years, filters, ounit,
  region, agg_region, add_global)
}
\description{
Produce oil production share by type by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{subsector}
  \item{technology}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
