% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module.pcgdp_ppp_}
\alias{module.pcgdp_ppp_}
\title{Per capita GDP PPP Data Module}
\usage{
module.pcgdp_ppp_(mode, allqueries, aggkeys, aggfn, strtyr, endyr, filters,
  ounit)
}
\description{
Produce GDP PPP by region.
}
\details{
Columns:
\itemize{
  \item{region}
  \item{year}
  \item{value}
  \item{units}
}
}
\keyword{internal}
