% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module.population_growth}
\alias{module.population_growth}
\title{Population Growth Data Module}
\usage{
module.population_growth(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce population growth by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
