% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_energy.R
\name{module.primary_energy_direct}
\alias{module.primary_energy_direct}
\title{Primary Energy (Direct Equivalent) Data Module}
\usage{
module.primary_energy_direct(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce primary energy by fuel
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
