% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_energy.R
\name{module.primary_energy_direct_pc}
\alias{module.primary_energy_direct_pc}
\title{Primary Energy (Direct Equivalent) Per Capita Data Module}
\usage{
module.primary_energy_direct_pc(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce primary energy per capita by fuel
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
