% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_energy.R
\name{module.resource_production_pc}
\alias{module.resource_production_pc}
\title{Resource Production Per Capita Data Module}
\usage{
module.resource_production_pc(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit, region, agg_region, add_global)
}
\description{
Produce resource production per capita by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{subsector}
  \item{technology}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
