% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{module.sales}
\alias{module.sales}
\title{Sales Data Module}
\usage{
module.sales(mode, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\description{
Produce sales
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
