% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{module.service_intensity}
\alias{module.service_intensity}
\title{Service Intensity Data Module}
\usage{
module.service_intensity(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Calculate service intensity by submode
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
