% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{module.test_fun}
\alias{module.test_fun}
\title{Test function for runModule}
\usage{
module.test_fun(mode, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\description{
This function prints a diagnostic and returns an empty data frame.
}
\keyword{internal}
