% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.total_enduse_intensity_ar4}
\alias{module.total_enduse_intensity_ar4}
\title{End-Use Sector GHG Emissions Intensity Data Module}
\usage{
module.total_enduse_intensity_ar4(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce total sector ghg emissions intensity - includes indirect emissions from electricity and biomass CO2 accounting.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
