% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{module.transportation_final_energy}
\alias{module.transportation_final_energy}
\title{Energy Data Module}
\usage{
module.transportation_final_energy(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce final energy by technology and vintage
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
