% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{module.transportation_service_output}
\alias{module.transportation_service_output}
\title{Service Output Data Module}
\usage{
module.transportation_service_output(mode, allqueries, aggkeys, aggfn, years,
  filters, ounit)
}
\description{
Produce service output by technology and vintage
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
