% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{module.water_consumption}
\alias{module.water_consumption}
\title{Water Consumption Module}
\usage{
module.water_consumption(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce agricultural production by crop type
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{output}
  \item{technology}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
