% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{module.water_withdrawals}
\alias{module.water_withdrawals}
\title{Water Withdrawals Module}
\usage{
module.water_withdrawals(mode, allqueries, aggkeys, aggfn, years, filters,
  ounit)
}
\description{
Produce water withdrawals by region, aggregate sector, sector, subsector, or
technology. The column `aggregate sector` is created by further aggregating sectors
into one of Biomass, Primary Energy, Irrigation, Manufacturing, Electricity,
Municipal, or Livestock.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{sector}
  \item{subsector}
  \item{technology}
  \item{input}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
