% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{nameparse}
\alias{nameparse}
\title{Separate a filename into a stem and extension.}
\usage{
nameparse(name)
}
\arguments{
\item{name}{Name to parse}
}
\description{
The stem is everything up to the last '.'.  The extension is everything after
that.  If the filename has no '.'s in it, the stem is the whole name, and the
extension is NA.
}
\keyword{internal}
