% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Output function for CSV format}
\usage{
output(rslts, dataformat, fileformat, dirname)
}
\arguments{
\item{rslts}{Results tables from \code{\link{generate}}.  This must be either
a list of data frames or a list of lists of data frames.}

\item{dataformat}{Indicator of data format:  If 'tabs', write to separate files; if 'merged'
write merged results to a single file.}

\item{fileformat}{File format for the output:  'CSV' or 'XLSX'}

\item{dirname}{Directory to write output file(s) into.}
}
\description{
Output function for CSV format
}
\keyword{internal}
