% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_csv}
\alias{output_csv}
\alias{output_xlsx}
\title{Output function for CSV format}
\usage{
output_csv(rslts, tabs)

output_xlsx(rslts, tabs)
}
\arguments{
\item{rslts}{Results tables from \code{\link{generate}}.}

\item{tabs}{Flag indicating whether variables should be written to separate
tabs/files.}
}
\description{
Output function for CSV format
}
\section{Functions}{
\itemize{
\item \code{output_xlsx}: Output function for XLSX format
}}

