% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{parse_sector}
\alias{parse_sector}
\title{Parse sector column}
\usage{
parse_sector(df, hasvintage, hasfuel, hastechnology)
}
\arguments{
\item{df}{Data returned for individual query}

\item{hasvintage}{Logical indicating if data has vintage col from query file}

\item{hasfuel}{Logical indicating if data has fuel and liquid_type col after fuel() processing}

\item{hastechnology}{Logical indication if data has technology col from query file}
}
\description{
Service, mode, and submode can be parsed from sector and subsector cols of
query data. Some sectors are disaggregated versions of other sectors, making
the data redundant. This is handled by filling service, mode, and submode cols
only for those observations that meet present demands of reporting templates.
}
\keyword{internal}
