% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{parse_yrlist}
\alias{parse_yrlist}
\title{Parse a string describing the years to include in the output}
\usage{
parse_yrlist(yrlist)
}
\arguments{
\item{yrlist}{String containing the year list}
}
\description{
Parse the input string and return an integer vector of years to include.
}
\details{
The years are specified as a comma-separated list, where each element is one
of:
\itemize{
  \item{A single integer, which includes that individual year}
  \item{A range in the form start:end, which includes all years from start to
end, inclusive.}
  \item{A stepped range in the form start:end:step, which includes all years
from start to end in steps of step.}
}
}
\keyword{internal}
