% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqueries.R
\name{parseQueries}
\alias{parseQueries}
\title{Parse queries from an XML query file}
\usage{
parseQueries(qfile)
}
\arguments{
\item{qfile}{File containing the XML definitions of the queries.}
}
\value{
\code{invisible(NULL)}
}
\description{
This function parses queries from the input file and stores them in the
package namespace, where \code{\link{runQueries}} can use them to run any
queries requested.
}
\details{
This function has the side effect of storing the queries retrieved from the
file.  The stored list will persist until the R session is restarted and the
package is reloaded.  Subsequent calls will accumulate in the list.  If the
titles of any new queries duplicate existing queries, the new ones will
overwrite the old ones.
}
