% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqueries.R
\docType{data}
\name{private}
\alias{private}
\title{Storage space for package-private variables}
\format{An object of class \code{environment} of length 1.}
\usage{
private
}
\description{
Storage space for package-private variables
}
\keyword{internal}
