% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{process_scenario}
\alias{process_scenario}
\title{Run queries and process results for a single scenario}
\usage{
process_scenario(scen, dbloc, dbname, q2run, varctl, regionctl, add_global)
}
\arguments{
\item{scen}{Scenario name}

\item{dbloc}{Directory containing the database}

\item{dbname}{Name of the database}

\item{q2run}{Character vector of titles of queries to run.}

\item{varctl}{Table read in from the variable control file.}
}
\description{
This is the main work function for \code{\link{generate}} and should only be
called from there.
}
\keyword{internal}
