% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_agg.R
\name{region_agg}
\alias{region_agg}
\title{Convert a table to aggregate regions}
\usage{
region_agg(tbl, region, agg_region)
}
\arguments{
\item{tbl}{Table to aggregate}

\item{region}{List of regions in tbl}

\item{agg_region}{List of regions to aggregate tbl to}
}
\description{
Aggregate the input table using the specified aggregation regions
}
\keyword{internal}
