% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{rename_length_1}
\alias{rename_length_1}
\title{Work function for generate climateworks format}
\usage{
rename_length_1(data, old_name, new_name)
}
\arguments{
\item{data}{Tibble of GCAM data}

\item{old_name}{Column name we are replacing}

\item{new_name}{New column name}
}
\description{
Work function for generate climateworks format
}
\keyword{internal}
