% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{rename_length_2}
\alias{rename_length_2}
\title{Work function for generate climateworks format}
\usage{
rename_length_2(data, old_names, new_names)
}
\arguments{
\item{data}{Tibble of GCAM data}

\item{old_names}{Vector of two column names we are replacing}

\item{new_names}{Vector of two new column names}
}
\description{
Work function for generate climateworks format
}
\keyword{internal}
