% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{runModule}
\alias{runModule}
\title{Run a data module}
\usage{
runModule(var, mode, allqueries = NULL, aggkeys = NULL, aggfn = NULL,
  strtyr = NULL, endyr = NULL, filters = NULL, ounit = NULL)
}
\arguments{
\item{var}{Name of the variable to produce.  This must be one of the varibles
known to the system.  The \code{\link{listModules}} function lists the known
variables.}

\item{mode}{Either \code{getq} or \code{run}.  The former returns
the names of the GCAM queries needed for the calculation; the latter runs the
calculation.  In \code{getq} mode all of the remaining arguments are
ignored.}

\item{allqueries}{List of all the queries pulled by the system.}

\item{aggkeys}{Character string listing the aggregation columns desired.
All of these columns must be present in the table computed by the main body
of the function.  If empty or NULL, no aggregation will be performed.}

\item{aggfn}{Function to use for aggregating.  If none specified,
\code{\link{base::sum}} will be used.}

\item{strtyr}{Start year.  All years prior to this will be dropped.}

\item{endyr}{End year.  All years after this will be dropped.}

\item{filters}{Character string giving a list of additional filters to be
applied, in s-exp format.}

\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
Run the module associated with object \code{var}.  If mode is \code{\link{getq}},
return a character vector of the names of all GCAM queries needed as input by
the module.  If mode is \code{\link{run}}, return a data frame containing the
processed variable.
}
