% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqueries.R
\name{runQueries}
\alias{runQueries}
\title{Run the input list of queries against a GCAM output database.}
\usage{
runQueries(qlist, dbloc, dbfile, scenario = NULL)
}
\arguments{
\item{qlist}{Character vector of titles of queries to run.}

\item{dbloc}{Name of the directory containing the database.}

\item{dbfile}{Name of the database}

\item{scenario}{Name of the scenario to query}
}
\description{
Run all of the queries in the input list against the specified database.  The
result will be returned as a list of data frames.
}
\details{
The \code{dbloc}, and \code{dbfile} arguments give the location and name of
the database.  The \code{scenario} argument is passed directly to
\code{\link[rgcam]{runQuery}}, and so has the same options as the
corresponding argument of that function.  Notably, passing \code{NULL} will
cause the last scenario in the database, irrespective of its name, to be
retrieved.

The input query list is a list of query \emph{titles}.  These will be looked
up in the list of query strings parsed by \code{\link{parseQueries}}.
Therefore, it is an error to call this function without first calling that
one.  It is similarly an error to include in the query list a query that
wasn't parsed in the call to \code{parseQueries}.
}
\keyword{internal}
