% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{semiaggregate}
\alias{semiaggregate}
\title{Semiaggregate}
\usage{
semiaggregate(df)
}
\arguments{
\item{df}{Data returned for individual query}
}
\description{
some queries return queries with disaggregated sectors/subsectors, which requires ones like (trn_pass_road_LDV_2W, moped)
}
\keyword{internal}
