% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{trans_filter_svc}
\alias{trans_filter_svc}
\title{Filter raw queries by passenger services}
\usage{
trans_filter_svc(qlist, passenger = TRUE)
}
\arguments{
\item{qlist}{List of queries relevant to the module at hand (i.e., pass only
the queries that the module will use.)}

\item{passenger}{If \code{TRUE}, filter to passenger services; if
\code{FALSE} filter to freight services.}
}
\description{
Filter raw queries by passenger services
}
\keyword{internal}
