% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{trans_standardize}
\alias{trans_standardize}
\title{Standardize transportation query dataframes}
\usage{
trans_standardize(queryData)
}
\arguments{
\item{queryData}{Data returned for one variable query}
}
\description{
Convert the GCAM representation using a hodgepodge of sectors, subsectors,
and technologies into the standard transportation taxonomy of service, mode,
submode, fuel, and vintage.
}
\details{
Columns removed (no mapping): rundate, load-factor
Columns removed (after mapping) : sector, subsector
Technologies removed: adv-electric, adv-liquid, tech-adv-electric, and
tech-adv-liquid (only appears in load factors)
Columns aggregated over : technology, vintage -- include in dplyr::group_by() if needed later

Final, standardized set of columns : Units, scenario, region, year, service,
mode, submode, fuel, liquid_type If this set changes, need to adjust call to
group_by() below All other cols are either dropped or aggregated over
}
\keyword{internal}
