% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitasserts}
\alias{unitasserts}
\title{Common assertions on unit string inputs}
\usage{
unitasserts(iunit, ounit)
}
\arguments{
\item{iunit}{Input unit}

\item{ounit}{Output unit}
}
\description{
All of the unit conversion functions are looking for a length-1 character
vector.  This function checks that these conditions are met.
}
\details{
The conditions are applied as assertions and will therefore cause errors
rather than warnings if they fail.  This a departure from our usual
practice.  We do it because having the wrong length or wrong type is a
programming error, not a data error.  Therefore, we expect these to trigger
only in during testing new modules, in which case an error and immediate halt
is appropriate.
}
