% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{unitconv}
\alias{unitconv}
\title{Unit conversion}
\usage{
unitconv(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts module data into output units.
}
\keyword{internal}
