% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_area}
\alias{unitconv_area}
\title{area conversion}
\usage{
unitconv_area(iunit, ounit, inverse = FALSE)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts area
}
\keyword{internal}
