% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_counts}
\alias{unitconv_counts}
\title{Counts}
\usage{
unitconv_counts(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts data reported in counts, e.g. thousands, millions, etc..
}
\keyword{internal}
