% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_energy}
\alias{unitconv_energy}
\title{Energy}
\usage{
unitconv_energy(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts data reported as energy, e.g. exaJoules,
MegaWatt-hours, etc.
}
\keyword{internal}
