% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_ldfctr}
\alias{unitconv_ldfctr}
\title{Load factor}
\usage{
unitconv_ldfctr(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
GCAM outputs this variable as load/veh, whether Passenger or Freight.
Passenger is usually reported as 'persons/vehicle' and Freight as 'tonnes/vehicle
Unit conversion functions are called after filtering and aggregation, meaning module_data
is a simple time series stripped of information identifying the type of variable reported (ie Passenger vs Freight)
Ounit must be specified correctly in the variable control file.
iunit is replaced with ounit silently
}
\keyword{internal}
