% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_mass}
\alias{unitconv_mass}
\title{weight conversion}
\usage{
unitconv_mass(iunit, ounit, inverse = FALSE)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts weight
}
\keyword{internal}
