% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_rupee}
\alias{unitconv_rupee}
\title{Unit conversion for USD into Rupee}
\usage{
unitconv_rupee(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
Convert to 2010 Rupees from USD1990
}
\keyword{internal}
