% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_usdollar}
\alias{unitconv_usdollar}
\title{Dollar Years}
\usage{
unitconv_usdollar(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts dollar values from different years
}
\keyword{internal}
