% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_vol}
\alias{unitconv_vol}
\title{volume conversion}
\usage{
unitconv_vol(iunit, ounit, inverse = FALSE)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts volume
}
\keyword{internal}
