% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_weight}
\alias{unitconv_weight}
\title{weight conversion}
\usage{
unitconv_weight(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
This function converts weight
}
\keyword{internal}
