% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{validate1}
\alias{validate1}
\title{Work function for validatectl}
\usage{
validate1(ctl, ctlname, expectcols, rqdcols)
}
\arguments{
\item{ctl}{Control file structure}

\item{ctlname}{Name of the control file we're testing (used in error messages)}

\item{expectcols}{Expected columns for this control file}

\item{rqdcols}{Columns for which data is required (no missing allowed)}
}
\description{
Work function for validatectl
}
\keyword{internal}
