% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{validatectl}
\alias{validatectl}
\title{Validate the structures read in from the control files}
\usage{
validatectl(scenctl, varctl)
}
\arguments{
\item{scenctl}{Scenario control file structure}

\item{varctl}{Variables control file structure}
}
\value{
NULL; Warnings or errors will be thrown as required.
}
\description{
Check for the following error conditions:
\itemize{
  \item{All expected columns are present.}
  \item{No empty or missing values in columns where missing values are not
permitted.}
}
Issue warnings for the following conditions:
\itemize{
  \item{Extraneous columns present}
}
}
\keyword{internal}
