% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{vint_tech_split}
\alias{vint_tech_split}
\title{Split technology into technology and vintage cols}
\usage{
vint_tech_split(df)
}
\arguments{
\item{df}{Data returned for individual query}
}
\description{
Query returns data with Technology column in the format'Technology, year=Vintage'.
This function returns the data frame with Technology split into a Technology and
Vintage column.
}
\keyword{internal}
