% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{yrparse}
\alias{yrparse}
\title{Parse year values out of a string}
\usage{
yrparse(yrstr)
}
\arguments{
\item{yrstr}{Character vector of strings to be parsed.}
}
\value{
Integer vector of years.
}
\description{
We don't know whether we will be getting two or 4 digit years, so we try to
figure it out as best we can.
}
\keyword{internal}
