% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate a table by specified keys}
\usage{
aggregate(tbl, aggfn, aggkeys)
}
\arguments{
\item{tbl}{Table to aggregate}

\item{aggfn}{String giving the aggregation function.  If missing, use
\code{\link[base]{sum}}.}

\item{aggkeys}{String giving a list of aggregation keys.  If missing, return
the table unmodified.}
}
\description{
Aggregate the input table using the specified aggregation function,
using the specified columns as keys.
}
\keyword{internal}
