% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{listVariables}
\alias{listVariables}
\title{List the variables that the system knows how to generate}
\usage{
listVariables()
}
\description{
Returns a character vector listing the variables that currently have recipes
in the system.
}
