% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transportation.R
\name{mapfuel}
\alias{mapfuel}
\title{Fuel}
\usage{
mapfuel(en, ref)
}
\arguments{
\item{en}{Data returned for final energy query}

\item{en}{Data returned for refined liquids query}
}
\description{
Service, mode, and submode can be parsed from sector and subsector cols of
query data. Some sectors are disaggregated versions of other sectors, making
the data redundant. This is handled by filling service, mode, and submode cols
only for those observations that meet present demands of reporting templates.
}
\keyword{internal}
