% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\name{module.co2_emissions}
\alias{module.co2_emissions}
\title{CO2 emissions Data Module}
\usage{
module.co2_emissions(mode, allqueries, aggkeys, aggfn, years, filters, ounit)
}
\description{
Produce service output by technology and vintage
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
