% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeconomics.R
\name{module.population}
\alias{module.population}
\title{Population Data Module}
\usage{
module.population(mode, allqueries, aggkeys, aggfn, strtyr, endyr, filters,
  ounit)
}
\description{
Produce population by region.
}
\details{
The raw table used by this module has columns:
\itemize{
  \item{scenario}
  \item{region}
  \item{year}
  \item{value}
  \item{Units}
}
}
\keyword{internal}
