% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_csv}
\alias{output_csv}
\alias{output_xlsx}
\title{Output function for CSV format}
\usage{
output_csv(rslts, dataformat, dirname)

output_xlsx(rslts, tabs, dirname)
}
\arguments{
\item{rslts}{Results tables from \code{\link{generate}}.  This must be either
a list of data frames or a list of lists of data frames.}

\item{dataformat}{Indicator of data format:  If 'tabs', write to separate files; if 'merged'
write merged results to a single file.}

\item{dirname}{Directory to write output file(s) into.}
}
\description{
Output function for CSV format
}
\section{Functions}{
\itemize{
\item \code{output_xlsx}: Output function for XLSX format
}}

\keyword{internal}
