% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqueries.R
\name{parseQueries}
\alias{parseQueries}
\title{Parse queries from an XML query file}
\usage{
parseQueries(qfile)
}
\arguments{
\item{qfile}{File containing the XML definitions of the queries.}
}
\value{
\code{invisible(NULL)}
}
\description{
Parses queries from the input file and store them in the
package namespace, where \code{\link{generate}} can use them to run any
queries requested.
}
\details{
This function has the side effect of storing the queries retrieved from the
file.  The stored list will persist until the R session is restarted and the
package is reloaded.  Subsequent calls will accumulate in the list.  If the
titles of any new queries duplicate existing queries, the new ones will
overwrite the old ones.

The function is run automatically at package startup, using the default query
definitions shipped with the package.  Therefore, users shouldn't ordinarily
need to run this function directly.  However, it may occasionally be useful
to redefine one of the existing queries, if, for example, you have output
databases that don't follow the normal schema for some reason.

In theory you could add queries beyond the ones in the default queries
configuration, but since none of the other code in the package will refer to
them, this is primarily useful only to package developers for testing new
processing code under development.
}
