% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\docType{data}
\name{RUN}
\alias{RUN}
\title{Token indicating run mode for \code{\link{runModule}} family of functions.}
\format{An object of class \code{character} of length 1.}
\usage{
RUN
}
\description{
Token indicating run mode for \code{\link{runModule}} family of functions.
}
\keyword{internal}
