% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv}
\alias{unitconv}
\alias{unitconv_counts}
\alias{unitconv_energy}
\alias{unitconv_usdollar}
\title{Calculate unit conversion factors}
\usage{
unitconv_counts(iunit, ounit, inverse = FALSE)

unitconv_energy(iunit, ounit, inverse = FALSE)

unitconv_usdollar(iunit, ounit, inverse = FALSE)
}
\arguments{
\item{iunit}{Units for the input data}

\item{ounit}{Desired output unit.}

\item{inverse}{If \code{TRUE} invert the conversion factor.  Use this in
cases where the unit being converted is in the denominator; \emph{e.g.}
cost/EJ -> cost/MWh.}
}
\description{
Return a factor for converting the input unit to the output unit.  This
factor should be applied as a multiplier:  out = in * fac.  In the event the
conversion could not be performed (e.g., because one of the units couldn't be
identified), \code{NA} will be returned (we return NA instead of NULL so that
you can multiply several conversion factors and check once at the end whether
they all succeeded).
}
\section{Functions}{
\itemize{
\item \code{unitconv_counts}: Convert counting units (thousands, millions, etc.)

Supported units are described in \code{\link{parsecounts}}.

\item \code{unitconv_energy}: Convert energy units

Supported units: EJ, TJ, MJ, MWh, TWh

\item \code{unitconv_usdollar}: Adjust US Dollar units for inflation.

Convert US dollar units from one dollar year to another using the GDP
deflator obtained from the Federal Reserve's FRED database.
}}

