% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unitconv_usdollar}
\alias{unitconv_usdollar}
\title{Unit conversion for dollar years}
\usage{
unitconv_usdollar(module_data, ounit)
}
\arguments{
\item{ounit}{Desired output unit.  If omitted, results will be returned with
no unit conversion.}
}
\description{
Convert US dollar units from one dollar year to another using the GDP
deflator obtained from the Federal Reserve's FRED database.
}
\keyword{internal}
